+++ src/behavior.rs
@@ -335,8 +335,12 @@ pub fn update_sand(sandbox: &mut Sandbox, x: usize, y: usize) {
 
 pub fn update_water(sandbox: &mut Sandbox, x: usize, y: usize) {
     if sandbox[x][y].unwrap().tempature >= 100 {
-        sandbox[x][y] = None;
-        return;
+        let t = sandbox[x][y].unwrap().tempature as f64 / 150.0;
+        let chance = (1.0 - t) * 0.3 + t * 0.7;
+        if sandbox.rng.gen_bool(chance) {
+            sandbox[x][y].as_mut().unwrap().ptype = ParticleType::Steam;
+            return;
+        }
     }
 
     // Find the first dry Sand below, delete this particle, and turn the Sand wet.
@@ -741,5 +745,11 @@ pub fn update_mirror(sandbox: &mut Sandbox, x: usize, y: usize) {
                 }
             }
         }
+    }
+}
+
+pub fn update_steam(sandbox: &mut Sandbox, x: usize, y: usize) {
+    if sandbox[x][y].unwrap().tempature < 100 {
+        sandbox[x][y].as_mut().unwrap().ptype = ParticleType::Water;
     }
 }