+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Performance improvements
 * Acid, Fire, and Electricity now have a glowing effect
 * Fixed a bug where stacking any particle on top of dry Sand would let Water go through that particle to turn the Sand wet
+* Added the Steam particle
 
 # 1.0.0 (2020-07-12)
 * First release